Update ft_print()
Update ft_printf.c ft_printf.h

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: ipersids <ipersids@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/11 19:02:08 by ipersids          #+#    #+#             */
-/*   Updated: 2024/11/13 23:41:32 by ipersids         ###   ########.fr       */
+/*   Updated: 2024/11/14 12:52:05 by ipersids         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,6 +22,6 @@ int		ft_printf(const char *format, ...);
 
 size_t	ft_putunbr_fd(unsigned int n, int fd);
 size_t	ft_puthex_fd(unsigned int n, int fd, short int is_uppercase);
-size_t	ft_putptr_fd(unsigned int n, int fd);
+size_t	ft_putptr_fd(unsigned long long n, int fd);
 
 #endif
